otpc/requirements.txt
biopython==1.85
numpy==2.2.2
pandas==2.2.3
pyfastx==2.2.0
pysam==0.23.0
python-dateutil==2.9.0.post0
pytz==2025.1
setuptools==75.8.0
six==1.17.0
tzdata==2025.1
